Updated join group page
